+++ Assets/Playground Scenes/SaveScreenShotFunctionality.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1fc4c9a062a794e84bb7bc4546508ed6
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
