+++ Assets/Prefabs/Props/Clutter Props/Print Out Variant.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1e4f1896ecc434d30b0ea72e0666860b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
