--- Assets/Scripts/TapeInformation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d99795795af38426e8ecb229b4a04519
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
