--- Assets/Scripts/TapeSOs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d6fcbf9738e8d4e779e56c4d87ffa228
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
