+++ Assets/Scripts/TapeScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fba1219a1845f440d8c1b22400c774bb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
