+++ Assets/Scripts/TapeScripts/PrintScreenshot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f7db3cab4ef874377847041afcd475f9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
