+++ Assets/Scripts/TapeScripts/TapeSOs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d6fcbf9738e8d4e779e56c4d87ffa228
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
