+++ Assets/Scripts/TapeScripts/TapeSOs/Tape1.asset
@@ -1,0 +1,20 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1292132138c7046129596273a38de46e, type: 3}
+  m_Name: Tape1
+  m_EditorClassIdentifier: 
+  originalCorruptedVideoClip: {fileID: 32900000, guid: cfdc14c08821d48b6a9c032661061a1c, type: 3}
+  branchACorruptedVideoClip: {fileID: 32900000, guid: b45f6932bec8647afb06d75b0c439ebb, type: 3}
+  branchBCorruptedVideoClip: {fileID: 32900000, guid: b5394e95cd25b40679d9467dd41aa24c, type: 3}
+  branchASolutionVideoClip: {fileID: 32900000, guid: 1afcd5133d42f4331beab3838fbcf7b8, type: 3}
+  branchBSolutionVideoClip: {fileID: 32900000, guid: b31de4387aad3416a8cb5a75a9c8ea91, type: 3}
+  timeGlitchFixedInFixedTape: 60
