+++ Assets/Scripts/TapeScripts/TapeSOs/Tape1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f28026c8e159d4a9ba643e88f3b2ca72
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
