+++ Assets/Scripts/TapeScripts/TapeSOs/TapeSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1292132138c7046129596273a38de46e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
