+++ Assets/Textures/New Material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c935d60ecb8404fa8a5e159e7d8fa37b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
